[refactor] DI: split CoilModule from AppModule for test doubles
